[DOCS] include `validate.py` in `gen.py`
